ui improvement
+ move sidebar to bottom on mobile
+ move signout button to profile page on mobile
+ add useLogout hook for logging out
+ minor improve on notification page(user now can't use delete all notifications button if there is no notif)

+++ frontend/src/pages/notification/NotificationPage.jsx
@@ -87,6 +87,10 @@ const NotificationPage = () => {
 
 	const deleteNotifications = () => {
 		if (isDeleteAllNotifsPending) return;
+		if (!notifications || notifications?.length === 0) {
+			toast.error("notification list is empty");
+			return;
+		}
 		deleteAllNotifications();
 	};
